lib: sbi_domain: Support Zcheripurecap ABI for hart context
Use uintptr_t for stvec, sscratch and sepc for Zcheripurecap ABI
compatible, as they are pointers rather than a plain integer address.

+++ lib/sbi/sbi_domain_context.c
@@ -29,11 +29,11 @@ struct hart_context {
 	/** Supervisor interrupt enable register */
 	unsigned long sie;
 	/** Supervisor trap vector base address register */
-	unsigned long stvec;
+	uintptr_t stvec;
 	/** Supervisor scratch register for temporary storage */
-	unsigned long sscratch;
+	uintptr_t sscratch;
 	/** Supervisor exception program counter register */
-	unsigned long sepc;
+	uintptr_t sepc;
 	/** Supervisor cause register */
 	unsigned long scause;
 	/** Supervisor trap value register */
